+++ CHANGELOG.md
@@ -12,8 +12,8 @@ This release removes the long-press event handling completely. It seems it can n
 
 - Feat: Improve playlist shuffle performance
 - Upd: Use signalfd
+- Upd: Improve websocket hanlding
 - Fix: Remove long-press events #1556
-- Fix: Set websocket connect timeout to 3s
 
 ***